Refactor Supernova benchmarks and config for CI
