+++ benches/supernova-ci.rs
@@ -1,9 +1,11 @@
 use criterion::*;
+
 use std::time::Duration;
 
 mod common;
 use common::supernova::targets::{
-  bench_one_augmented_circuit_recursive_snark, bench_two_augmented_circuit_recursive_snark,
+  bench_two_augmented_circuit_compressed_snark_with_computational_commitments,
+  bench_two_augmented_circuit_recursive_snark,
 };
 
 // To run these benchmarks, first download `criterion` with `cargo install cargo-criterion`.
@@ -13,17 +15,17 @@ use common::supernova::targets::{
 cfg_if::cfg_if! {
   if #[cfg(feature = "flamegraph")] {
     criterion_group! {
-      name = recursive_snark_supernova;
+      name = supernova_ci;
       config = Criterion::default().warm_up_time(Duration::from_millis(3000)).with_profiler(pprof::criterion::PProfProfiler::new(100, pprof::criterion::Output::Flamegraph(None)));
-      targets = bench_one_augmented_circuit_recursive_snark, bench_two_augmented_circuit_recursive_snark
+      targets = bench_two_augmented_circuit_recursive_snark, bench_two_augmented_circuit_compressed_snark_with_computational_commitments
     }
   } else {
     criterion_group! {
-      name = recursive_snark_supernova;
+      name = supernova_ci;
       config = Criterion::default().warm_up_time(Duration::from_millis(3000));
-      targets = bench_one_augmented_circuit_recursive_snark, bench_two_augmented_circuit_recursive_snark
+      targets = bench_two_augmented_circuit_recursive_snark, bench_two_augmented_circuit_compressed_snark_with_computational_commitments
     }
   }
 }
 
-criterion_main!(recursive_snark_supernova);
+criterion_main!(supernova_ci);
